pyproject.toml
[project]
name = "auto-reading-lib"
version = "0.1.0"
description = "Shared library for auto-reading Claude Code Skills"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib"]

[tool.pytest.ini_options]
testpaths = ["tests"]
